requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
